+++ src/form/b-checkbox/b-checkbox.ts
@@ -26,7 +26,6 @@ import iInput, {
 	prop,
 	system,
 	computed,
-	p,
 
 	ModsDecl,
 	ModEvent,
@@ -134,7 +133,6 @@ export default class bCheckbox extends iInput implements iSize {
 	readonly changeable: boolean = true;
 
 	/** @override */
-	@p({replace: false})
 	get value(): this['Value'] {
 		const
 			{checked} = this.mods;
@@ -304,7 +302,6 @@ export default class bCheckbox extends iInput implements iSize {
 	}
 
 	/** @override */
-	@p({replace: false})
 	protected resolveValue(value?: this['Value']): this['Value'] {
 		const
 			i = this.instance;