+++ src/super/i-block/interface.ts
@@ -120,12 +120,6 @@ export interface UnsafeIBlock<CTX extends iBlock = iBlock> extends UnsafeCompone
 	ifOnceStore: CTX['ifOnceStore'];
 
 	// @ts-ignore (access)
-	activated: CTX['activated'];
-
-	// @ts-ignore (access)
-	deactivated: CTX['deactivated'];
-
-	// @ts-ignore (access)
 	syncRouterState: CTX['syncRouterState'];
 
 	// @ts-ignore (access)